Remove double bracket array shorthand
This is a breaking change. The `[[...array]]` has been dropped. Instead
an `array` function is added to the env by default. This function can be
overwritten if desired.

Migration: change `[[...array]]` to `['array', ...array]`.

The main reason for this change is to support having the operator be an
anonymous function. e.g. `[[add, 1], 2]`. Previously this would result
in an array as the partially applied `add(1)` is not JSON serializable.